Move application_status_update to its own form object for now
