+++ app/views/forms/_application_status_form.html.erb
@@ -1,26 +1,38 @@
 <%= turbo_frame_tag "participant_application_status_form_#{participant.id}" do %>
   <p>
-    <%= participant.name %>
+    Applicant Name: <%= participant.name %>
   </p>
 
-  <% participant.application_question_responses.each do |prompt, response|%>
+  <% participant.application_question_responses.each do |application_question_response|%>
     <p>
-      <%= prompt %>
+      Question: <%= application_question_response.application_question.prompt %>
     <br>
-      <%= response %>
+      <%= application_question_response.value %>
     </p>
   <% end %>
 
-  <%= simple_form_for participant do |f| %>
+  <%= simple_form_for :application_status_form, url: application_status_forms_path do |f| %>
+    <%= f.hidden_field :participant_id, value: participant.id %>
+    <%= f.hidden_field :workshop_id, value: participant.workshop_id %>
     <%= f.hidden_field :application_status, value: "accepted" %>
-    <%= f.submit "Accept participant application", class: "text-white bg-green-700 hover:bg-green-800 focus:outline-none focus:ring-4 focus:ring-green-300 font-medium rounded-full text-sm px-5 py-2.5 text-center me-2 mb-2" %>
+    <%= f.submit "Accept participant application", class: "text-white bg-green-700 hover:bg-green-800 
+                                                           focus:outline-none focus:ring-4 focus:ring-green-300 
+                                                           font-medium rounded-full text-sm px-5 py-2.5 text-center 
+                                                           me-2 mb-2" %>
   <% end %>
-  <%= simple_form_for participant do |f| %>
+  <%= simple_form_for :application_status_form, url: application_status_forms_path do |f| %>
+    <%= f.hidden_field :participant_id, value: participant.id %>
+    <%= f.hidden_field :workshop_id, value: participant.workshop_id %>
     <%= f.hidden_field :application_status, value: "waitlisted" %>
-    <%= f.submit "Waitlist participant application", class: "text-white bg-yellow-400 hover:bg-yellow-500 focus:outline-none focus:ring-4 focus:ring-yellow-300 font-medium rounded-full text-sm px-5 py-2.5 text-center me-2 mb-2 " %>
+    <%= f.submit "Waitlist participant application", class: "text-white bg-yellow-400 hover:bg-yellow-500 
+                                                             focus:outline-none focus:ring-4 focus:ring-yellow-300 font-medium rounded-full 
+                                                             text-sm px-5 py-2.5 text-center me-2 mb-2 " %>
   <% end %>
-  <%= simple_form_for participant do |f| %>
+  <%= simple_form_for :application_status_form, url: application_status_forms_path do |f| %>
+    <%= f.hidden_field :participant_id, value: participant.id %>
+    <%= f.hidden_field :workshop_id, value: participant.workshop_id %>
     <%= f.hidden_field :application_status, value: "rejected" %>
-    <%= f.submit "Reject participant application", class: "text-white bg-red-700 hover:bg-red-800 focus:outline-none focus:ring-4 focus:ring-red-300 font-medium rounded-full text-sm px-5 py-2.5 text-center me-2 mb-2" %>
+    <%= f.submit "Reject participant application", class: "text-white bg-red-700 hover:bg-red-800 focus:outline-none 
+                                                           focus:ring-4 focus:ring-red-300 font-medium rounded-full text-sm px-5 py-2.5 text-center me-2 mb-2" %>
   <% end %>
 <% end %>
